+++ src/kinmu_lib/types/mod.rs
@@ -129,16 +129,16 @@ pub enum ScoreProp {
     PatternFixed((CondWrapper, Vec<Shift>, Score)),
     PatternGeneralAny((CondWrapper, Vec<Vec<Shift>>, Score)),
     PatternFixedAny((CondWrapper, Vec<Shift>, Score)),
-    Streak((CondWrapper, Vec<Shift>, isize, Score)),
+    Streak((CondWrapper, Vec<Shift>, i32, Score)),
     ShiftsBalance((CondWrapper, Shift, Shift, Score)),
     ShiftHalfBalance((CondWrapper, Shift, Score)),
     ShiftDirPriority((CondWrapper, Shift, Score)),
     DayCountRegardStaffAttribute((CondWrapper, Shift, StaffAttributeName, Score)),
     StaffCountRegardDayAttribute((CondWrapper, Shift, DayAttributeName, Score)),
-    StaffCount((CondWrapper, Shift, isize, Score)),
-    StaffCountWithPremise((CondWrapper, Shift, isize, CondWrapper, Shift, isize, Score)),
+    StaffCount((CondWrapper, Shift, i32, Score)),
+    StaffCountWithPremise((CondWrapper, Shift, i32, CondWrapper, Shift, i32, Score)),
     NGPair((CondWrapper, Shift, Score)),
-    NoSamePair((CondWrapper, isize, Shift, Score)),
+    NoSamePair((CondWrapper, i32, Shift, Score)),
 }
 
 impl fmt::Display for ScoreProp {
